+++ src/conn_interface.c
@@ -64,6 +64,7 @@ pthread_t crud_tid;
 /*----------------------------------------------------------------------------*/
 void timespec_diff(struct timespec *start, struct timespec *stop,
                    struct timespec *result);
+
 /*----------------------------------------------------------------------------*/
 /*                             External Functions                             */
 /*----------------------------------------------------------------------------*/
@@ -105,7 +106,7 @@ void createSocketConnection(void (* initKeypress)())
     StartThread(processUpstreamMessage, &upstream_msg_tid);
     ParodusMsgQ = NULL;
     StartThread(messageHandlerTask, &downstream_tid);
-    /*StartThread(serviceAliveTask, &svc_alive_tid;);*/
+    StartThread(serviceAliveTask, &svc_alive_tid);
     StartThread(CRUDHandlerTask, &crud_tid);
 
     if (NULL != initKeypress) 
@@ -167,25 +168,22 @@ void createSocketConnection(void (* initKeypress)())
             ParodusInfo("cloud_status set as %s after connection close\n", get_parodus_cfg()->cloud_status);
             if(get_parodus_cfg()->cloud_disconnect !=NULL)
             {
-				ParodusPrint("get_parodus_cfg()->cloud_disconnect is %s\n", get_parodus_cfg()->cloud_disconnect);
-				set_cloud_disconnect_time(CLOUD_RECONNECT_TIME);
-				ParodusInfo("Waiting for %d minutes for reconnecting .. \n", get_cloud_disconnect_time());
+                ParodusPrint("get_parodus_cfg()->cloud_disconnect is %s\n", get_parodus_cfg()->cloud_disconnect);
+		set_cloud_disconnect_time(CLOUD_RECONNECT_TIME);
+		ParodusInfo("Waiting for %d minutes for reconnecting .. \n", get_cloud_disconnect_time());
 
-				sleep( get_cloud_disconnect_time() * 60 );
-				ParodusInfo("cloud-disconnect reason reset after %d minutes\n", get_cloud_disconnect_time());
-				free(get_parodus_cfg()->cloud_disconnect);
-				reset_cloud_disconnect_reason(get_parodus_cfg());
+		sleep (get_cloud_disconnect_time() * 60);
+		ParodusInfo("cloud-disconnect reason reset after %d minutes\n", get_cloud_disconnect_time());
+		free(get_parodus_cfg()->cloud_disconnect);
+		reset_cloud_disconnect_reason(get_parodus_cfg());
             }
             createNopollConnection(ctx);
         }
-        clock_gettime(CLOCK_REALTIME, &stop);
-        timespec_diff(&start_svc_alive_timer, &stop, &diff);
-        if (diff.tv_sec >= 30) {
-                serviceAliveTask ();
-                start_svc_alive_timer.tv_sec += 30;
-        }
        } while(!get_close_retry() && !g_shutdown);
 
+    pthread_mutex_lock (get_global_svc_mut());
+    pthread_cond_signal (get_global_svc_con());
+    pthread_mutex_unlock (get_global_svc_mut());
     pthread_mutex_lock (get_global_crud_mut());
     pthread_cond_signal (get_global_crud_con());
     pthread_mutex_unlock (get_global_crud_mut());
@@ -197,6 +195,7 @@ void createSocketConnection(void (* initKeypress)())
     pthread_mutex_unlock (get_global_nano_mut());
 
     ParodusInfo ("joining threads\n");
+    JoinThread (svc_alive_tid);
     JoinThread (upstream_tid);
     JoinThread (downstream_tid);
     JoinThread (upstream_msg_tid);