+++ src/service_alive.h
@@ -24,6 +24,8 @@
 #ifndef _SERVICE_ALIVE_H_
 #define _SERVICE_ALIVE_H_
 
+#include <pthread.h>
+
 /*----------------------------------------------------------------------------*/
 /*                             Function Prototypes                            */
 /*----------------------------------------------------------------------------*/
@@ -32,8 +34,9 @@
 extern "C" {
 #endif
   
-int serviceAliveTask();
-
+void *serviceAliveTask();
+pthread_cond_t *get_global_svc_con(void);
+pthread_mutex_t *get_global_svc_mut(void);
 
 #ifdef __cplusplus
 }